+++ src/instance.ts
@@ -6,6 +6,7 @@ import { toUnix } from './helpers';
 import { Checker } from './checker';
 import { CompilerInfo, LoaderConfig, TsConfig } from './interfaces';
 import { WatchModeSymbol } from './watch-mode';
+import { createHash } from 'crypto';
 
 let colors = require('colors/safe');
 let pkg = require('../package.json');
@@ -19,8 +20,12 @@ export interface Instance {
     compilerConfig: TsConfig;
     loaderConfig: LoaderConfig;
     checker: Checker;
-    cacheIdentifier: any;
+    cacheIdentifier: string;
     context: string;
+
+    times: Dict<number>;
+    watchedFiles?: Set<string>;
+    startTime?: number;
 }
 
 export interface Compiler {
@@ -43,6 +48,7 @@ export interface Loader {
     emitFile: (fileName: string, text: string) => void;
     emitWarning: (msg: string) => void;
     emitError: (msg: string) => void;
+    context: string;
     options: {
         ts?: LoaderConfig
     };
@@ -58,7 +64,7 @@ export function getRootCompiler(compiler) {
     }
 }
 
-function resolveInstance(compiler, instanceName) {
+function resolveInstance(compiler, instanceName): Instance {
      if (!compiler._tsInstances) {
         compiler._tsInstances = {};
     }
@@ -87,7 +93,7 @@ export function ensureInstance(
     }
 
     const watching = isWatching(rootCompiler);
-    const context = process.cwd();
+    const context = options.context || process.cwd();
 
     let compilerInfo = setupTs(query.compiler);
     let { tsImpl } = compilerInfo;
@@ -114,6 +120,7 @@ export function ensureInstance(
 
     let babelImpl = setupBabel(loaderConfig, context);
     let cacheIdentifier = setupCache(
+        compilerConfig,
         loaderConfig,
         tsImpl,
         webpack,
@@ -144,7 +151,8 @@ export function ensureInstance(
         compilerConfig,
         checker,
         cacheIdentifier,
-        context
+        context,
+        times: {}
     };
 }
 
@@ -184,13 +192,13 @@ export function setupTs(compiler: string): CompilerInfo {
 }
 
 function setupCache(
+    compilerConfig: TsConfig,
     loaderConfig: LoaderConfig,
     tsImpl: typeof ts,
     webpack: Loader,
     babelImpl: any,
     context: string
-) {
-    let cacheIdentifier = null;
+): string {
     if (loaderConfig.useCache) {
         if (!loaderConfig.cacheDirectory) {
             loaderConfig.cacheDirectory = path.join(context, '.awcache');
@@ -200,14 +208,19 @@ function setupCache(
             mkdirp.sync(loaderConfig.cacheDirectory);
         }
 
-        cacheIdentifier = {
-            'typescript': tsImpl.version,
+        let hash = createHash('sha512') as any;
+        let contents = JSON.stringify({
+            typescript: tsImpl.version,
             'awesome-typescript-loader': pkg.version,
-            'awesome-typescript-loader-query': webpack.query,
-            'babel-core': babelImpl
-                ? babelImpl.version
-                : null
-        };
+            'babel-core': babelImpl ? babelImpl.version : null,
+            babelPkg: pkg.babel,
+            // TODO: babelrc.json/babelrc.js
+            compilerConfig,
+            env: process.env.BABEL_ENV || process.env.NODE_ENV || 'development'
+        });
+
+        hash.end(contents);
+        return hash.read().toString('hex');
     }
 }
 
@@ -218,7 +231,7 @@ function setupBabel(loaderConfig: LoaderConfig, context: string): any {
             let babelPath = loaderConfig.babelCore || path.join(context, 'node_modules', 'babel-core');
             babelImpl = require(babelPath);
         } catch (e) {
-            console.error(BABEL_ERROR);
+            console.error(BABEL_ERROR, e);
             process.exit(1);
         }
     }
@@ -328,6 +341,18 @@ export function readConfigFile(
 }
 
 let EXTENSIONS = /\.tsx?$|\.jsx?$/;
+export type Dict<T> = {[key: string]: T};
+
+const filterMtimes = (mtimes: any) => {
+    const res = {};
+    Object.keys(mtimes).forEach(fileName => {
+        if (!!EXTENSIONS.test(fileName)) {
+            res[fileName] = mtimes[fileName];
+        }
+    });
+
+    return res;
+};
 
 function setupWatchRun(compiler, instanceName: string) {
     compiler.plugin('watch-run', function (watching, callback) {
@@ -336,20 +361,50 @@ function setupWatchRun(compiler, instanceName: string) {
         const watcher = watching.compiler.watchFileSystem.watcher
             || watching.compiler.watchFileSystem.wfs.watcher;
 
-        const mtimes = watcher.mtimes || (watcher.getTimes && watcher.getTimes()) || {};
-        const changedFiles = Object.keys(mtimes).map(toUnix);
-        const updates = changedFiles
-            .filter(file => EXTENSIONS.test(file))
-            .map(changedFile => {
-                if (fs.existsSync(changedFile)) {
-                    checker.updateFile(changedFile, fs.readFileSync(changedFile).toString(), true);
+        const startTime = instance.startTime || watching.startTime;
+        const times = filterMtimes(watcher.getTimes());
+        const lastCompiled = instance.compiledFiles;
+
+        instance.compiledFiles = {};
+        instance.startTime = startTime;
+
+        const set = new Set(Object.keys(times).map(toUnix));
+        if (instance.watchedFiles || lastCompiled) {
+            const removedFiles = [];
+            const checkFiles = (instance.watchedFiles || Object.keys(lastCompiled)) as any;
+            checkFiles.forEach(file => {
+                if (!set.has(file)) {
+                    removedFiles.push(file);
+                }
+            });
+
+            removedFiles.forEach(file => {
+                checker.removeFile(file);
+            });
+        }
+        instance.watchedFiles = set;
+
+        const instanceTimes = instance.times;
+        instance.times = Object.assign({}, times) as any;
+
+        const updates = Object.keys(times)
+            .filter(fileName => {
+                const updated = times[fileName] > (instanceTimes[fileName] || startTime);
+                return updated;
+            })
+            .map(fileName => {
+                const unixFileName = toUnix(fileName);
+                if (fs.existsSync(unixFileName)) {
+                    checker.updateFile(unixFileName, fs.readFileSync(unixFileName).toString(), true);
                 } else {
-                    checker.removeFile(changedFile);
+                    checker.removeFile(unixFileName);
                 }
             });
 
         Promise.all(updates)
-            .then(() => callback())
+            .then(() => {
+                callback();
+            })
             .catch(callback);
     });
 }
@@ -385,11 +440,6 @@ function setupAfterCompile(compiler, instanceName, forkChecker = false) {
         const asyncErrors = watchMode === WatchMode.Enabled && !silent;
 
         let emitError = (msg) => {
-            if (compilation.bail) {
-                console.error('Error in bail mode:', msg);
-                process.exit(1);
-            }
-
             if (asyncErrors) {
                 console.log(msg, '\n');
             } else {
@@ -397,16 +447,32 @@ function setupAfterCompile(compiler, instanceName, forkChecker = false) {
             }
         };
 
-        instance.compiledFiles = {};
         const files = instance.checker.getFiles()
             .then(({files}) => {
-                Array.prototype.push.apply(compilation.fileDependencies, files.map(path.normalize));
+                const normalized = files.map(file => {
+                    const rpath = path.normalize(file);
+                    instance.compiledFiles[file] = true;
+                    return rpath;
+                });
+                Array.prototype.push.apply(compilation.fileDependencies, normalized);
             });
 
+        const timeStart = +(new Date());
         const diag = instance.loaderConfig.transpileOnly
             ? Promise.resolve()
             : instance.checker.getDiagnostics()
                 .then(diags => {
+                    if (!silent) {
+                        if (diags.length) {
+                            console.error(colors.red(`\n[${instanceName}] Checking finished with ${diags.length} errors`));
+                        } else {
+                            let timeEnd = +new Date();
+                            console.log(
+                                colors.green(`\n[${instanceName}] Ok, ${(timeEnd - timeStart) / 1000} sec.`)
+                            );
+                        }
+                    }
+
                     diags.forEach(diag => emitError(diag.pretty));
                 });